% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method}
\usage{
\method{coef}{linreg}(X)
}
\arguments{
\item{X}{An object}
}
\value{
A named vector of coefficients
}
\description{
Coefficients method
}
