\name{lab00004-package}
\alias{lab00004-package}
\alias{lab00004}
\docType{package}
\title{
\packageTitle{lab00004}
}
\description{
\packageDescription{lab00004}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab00004}
\packageIndices{lab00004}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{lab00004}

Maintainer: \packageMaintainer{lab00004}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation
directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
simple examples of the most important functions 
}
