% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression using Ordinary Least Squares}
\usage{
linreg(formula, data)
}
\arguments{
\item{X}{The design matrix (n x p) with n observations and p predictors.}

\item{y}{The response vector (n x 1) with n observations.}
}
\value{
An object of class "linreg" containing many parameters.
}
\description{
Linear Regression using Ordinary Least Squares
}
