% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{print}
\usage{
\method{print}{linreg}(X, ...)
}
\arguments{
\item{X}{}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An object
}
\description{
print
}
