% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression using Ordinary Least Squares}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{The formula.}

\item{data}{The data.}
}
\value{
An object of class "linreg" containing many parameters.
}
\description{
Linear Regression using Ordinary Least Squares
}
