% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Residuals vs Fitted and Scale-Location}
\arguments{
\item{model}{A linear model object created by lm().}
}
\description{
This function generates two diagnostic plots for linear regression models.
}
