% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.linreg.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residuals method}
\usage{
\method{resid}{linreg}(X, ...)
}
\arguments{
\item{X}{An object}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A vector of residuals
}
\description{
Residuals method
}
